refactor(terminal): extract shared BoxDetail component
